requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
python-dotenv
streamlit
requests
python-multipart
